\name{calibrator-package}
\alias{calibrator-package}
\alias{calibrator}
\docType{package}
\title{
\packageTitle{calibrator}
}
\description{
\packageDescription{calibrator}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{calibrator}
\packageIndices{calibrator}

}
\author{
\packageAuthor{calibrator}

Maintainer: \packageMaintainer{calibrator}
}
\references{
  \itemize{
    \item
    M. C. Kennedy and A. O'Hagan 2001. \dQuote{Bayesian calibration of
      computer models}.  Journal of the Royal Statistical Society, Series B,
    63(3): 425--464
    \item
    R. K. S. Hankin 2005. \dQuote{Introducing BACCO, an R bundle for
      Bayesian analysis of computer code output}, Journal of Statistical
    Software, 14(16)
  }
}
\keyword{ package }
