% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calMVA.R
\name{calMVA}
\alias{calMVA}
\title{Calibration of seasonal climate forecasts.}
\usage{
calMVA(y, x, newdata = NULL, crossval = TRUE, parallel = FALSE,
  max.ncores = 16, ncores = NULL)
}
\arguments{
\item{y}{climate4R grid. Reference observations the forecasts are calibrated towards (typically on a monthly/seasonal basis).}

\item{x}{climate4R grid. Forecasts used as predictors (typically a hindcast on a monthly/seasonal basis).}

\item{newdata}{Data to be corrected using the parameters from \code{x}. Typically the operational forecast.
By default \code{newdata} is \code{NULL}, so the hindcast (\code{x}) is corrected.}

\item{crossval}{Logical. Default to \code{TRUE}, performs a leave-one-year-out cross-validation, so
the parameters used for the calibration are computed using all years but the current one.
Unused when an operational forecast is passed to \code{newdata}.}

\item{parallel}{Logical. Should parallel execution be used?}

\item{max.ncores}{Integer. Upper bound for user-defined number of cores.}

\item{ncores}{Integer number of cores used in parallel computation. Self-selected number of
cores is used when \code{ncpus = NULL} (the default), or when \code{maxcores} exceeds the default \code{ncores} value.}
}
\value{
climate4R grid. Calibrated forecasts.
}
\description{
This function implements the method 1 described in Torralba et al. 2017, which has been previously applied 
in the context of seasonal forecasting to correct temperature and precipitation in Leung et al. 1999.
The ensemble mean produced by this method has the same mean and standard deviation as the observations. 
It is based on the assumption that both the reference and predicted distributions are approximated well by a Gaussian (normal) distribution.
}
\section{Parallel Processing}{


Parallel processing is enabled using the \pkg{parallel} package. 
Parallelization is undertaken by a FORK-type parallel socket cluster formed by \code{ncores}.
If \code{ncores} is not specified (default), \code{ncores} will be one less than the autodetected number of cores.
The maximum number of cores used for parallel processing can be set with the \code{max.ncores} argument, 
although this will be reset to the auto-detected number of cores minus 1 if this number is exceeded. Note that not all 
code, but just some critical loops within the function are parallelized.

In practice, parallelization does not always result in smaller execution times, due to the parallel overhead.
However, parallel computing may potentially provide a significant speedup for the 
particular case of large multimember datasets or large grids.
 
Parallel computing is currently not available for Windows machines.
}

\examples{
\dontrun{
## loading seasonal forecasts (CFS) and observations (NCEP) of boreal winter temperature over Iberia
data("CFS_Iberia_tas"); fcst = CFS_Iberia_tas
data("NCEP_Iberia_tas"); obs = NCEP_Iberia_tas
## passing from daily data to seasonal averages
fcst = aggregateGrid(fcst, aggr.y = list(FUN = "mean", na.rm = TRUE))
obs = aggregateGrid(obs, aggr.y = list(FUN = "mean", na.rm = TRUE))
## interpolating forecasts to the observations' resolution
fcst = interpGrid(fcst, new.coordinates = getGrid(obs))
## applying calibration
fcst.cal = calMVA(fcst, obs, crossval = TRUE)
## plotting climatologies
library(visualizeR)
spatialPlot(makeMultiGrid(climatology(obs),
                         climatology(fcst, by.member = FALSE),
                         climatology(fcst.cal, by.member = FALSE)),
           backdrop.theme = "coastline",
           layout = c(3, 1),
           names.attr = c("NCEP", "CFS (raw)", "CFS (calibrated)"))
}
}
\references{
\itemize{
\item Torralba, V., F.J. Doblas-Reyes, D. MacLeod, I. Christel, and M. Davis, 2017: Seasonal Climate Prediction: A New Source of Information for the Management of Wind Energy Resources. J. Appl. Meteor. Climatol., 56, 1231–1247, https://doi.org/10.1175/JAMC-D-16-0204.1 
\item Leung, L.R., A.F. Hamlet, D.P. Lettenmaier, and A. Kumar, 1999: Simulations of the ENSO hydroclimate signals in the Pacific Northwest Columbia River basin. Bull. Amer. Meteor. Soc., 80, 2313–2329, doi:10.1175/1520-0477(1999)080,2313: SOTEHS.2.0.CO;2
}
}
\seealso{
Other calibration: \code{\link{calCCR}},
  \code{\link{calLM}}, \code{\link{calNGR}},
  \code{\link{calRPC}}
}
\author{
R. Manzanas and V. Torralba.
}
