% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\name{cfs.forecast}
\alias{cfs.forecast}
\title{CFSv2 operational forecast}
\source{
All data have been retrieved from the User Data Gateway using the R package \pkg{loadeR.ECOMS} (Cofiño et al. 2018).
 The data structure is according to the climate4R common data model.
}
\description{
CFSv2 operational forecast (Nov 2015) for DJFMA 2016. See Details.
}
\details{
All data correspond to the ENSO 3.4 region, for the variable 2m air temperature, and have been interpolated
 onto de NCEP regular grid (2.5º hor. res.). The season is boreal winter (DJF), and the hindcast period is 1983-2009.
}
\references{
\itemize{
\item Cofiño, A.S., et al., 2018. The ECOMS User Data Gateway: Towards seasonal forecast data provision and research reproducibility
 in the era of Climate Services. Climate Services 9, 33–43. https://doi.org/10.1016/j.cliser.2017.07
\item climate4R web site: \url{http://www.meteo.unican.es/climate4R} 
}
}
