% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{corfun}
\alias{corfun}
\title{Ensemble Correlation}
\usage{
corfun(ens, obs, ...)
}
\arguments{
\item{ens}{ens}

\item{obs}{obs}

\item{...}{further argumentspassed to \code{\link[stats]{cor.test}}}
}
\description{
Ensemble correlation helper
}
\details{
For internal usage only, although exported for convenience
}
\keyword{internal}
