% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getPooledMemberStat}
\alias{getPooledMemberStat}
\title{Get pooled-member local statistics}
\usage{
getPooledMemberStat(grid, fun, ...)
}
\arguments{
\item{grid}{A climate4R CDM structure}

\item{fun}{Character string. Currently in use \code{"mean"} or \code{"sd"}, but any other valid function may work.}

\item{...}{Further args passed to \code{fun}}
}
\value{
A vector, in the appropriate coordinates ordering (consistent with \code{\link{array3Dto2Dmat}})
}
\description{
Retrieve statistics for each grid point by pooling together all members
}
\keyword{internal}
