% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getPooledMemberStat}
\alias{getPooledMemberStat}
\title{Get pooled-member local statistics}
\usage{
getPooledMemberStat(grid, fun, dim.along = "latlon", ...)
}
\arguments{
\item{grid}{A climate4R CDM structure}

\item{fun}{Character string. Currently in use \code{"mean"} or \code{"sd"}, but any other valid function may work.}

\item{dim.along}{Dimension to pool along. Default to \code{"latlon"}. Otherwise \code{"time"}}

\item{...}{Further args passed to \code{fun}}
}
\value{
A vector, in the appropriate coordinates ordering (consistent with \code{\link[transformeR]{array3Dto2Dmat}} in the case of
spatial aggregations or \code{\link[transformeR]{getRefDates}} in the case of time aggregations).
}
\description{
Retrieve statistics for each grid point by pooling together all members
}
\keyword{internal}
